apps/backend/pyproject.toml
[project]
name = "eduparent-backend"
version = "1.0.0"
description = "FastAPI backend for EduParent app"
authors = [
    {name = "EduParent Team"}
]
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.0.0",
    "sqlalchemy>=2.0.0",
    "python-multipart>=0.0.6",
    "alembic>=1.16.5",
    "psycopg2-binary>=2.9.10",
    "bcrypt>=4.3.0",
    "python-jose[cryptography]>=3.5.0",
    "email-validator>=2.3.0",
    "pydantic-ai>=0.0.14",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "httpx>=0.25.0",
    "pytest-asyncio>=0.21.0",
]
